requirements.txt
aiogram==3.7.0
gspread==6.1.2
google-auth==2.29.0
APScheduler==3.10.4
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
httpx==0.27.0
openpyxl==3.1.2
aiohttp==3.9.5
